backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
bcrypt==5.0.0
certifi==2025.8.3
cffi==2.0.0
click==8.2.1
cryptography==46.0.4
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.116.1
fastapi-cli==0.0.11
fastapi-cloud-cli==0.1.5
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
passlib==1.7.4
pyasn1==0.6.2
pycparser==3.0
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pymongo==4.15.0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rsa==4.9.1
sentry-sdk==2.37.1
shellingham==1.5.4
six==1.17.0
-e git+ssh://git@github.com/zoaient/Animest-in-love-revamp.git@c2d787cdfb71c0121a194673e9bcfc8a208b3b15#egg=slash_zoaient_backend&subdirectory=backend
sniffio==1.3.1
starlette==0.47.3
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
-e git+ssh://git@github.com/zoaient/Animest-in-love-revamp.git@c2d787cdfb71c0121a194673e9bcfc8a208b3b15#egg=visual_novel_backend&subdirectory=backend
watchfiles==1.1.0
websockets==15.0.1
